fixed minor Wall errors in 01 06 07

+++ 2021/include/01.hpp
@@ -22,7 +22,7 @@ void Sol01<T>::Solution1() {
     obj = ReadVVT<T>(this->filename, ',');
     std::vector<T> out = vTFromVVT<T>(obj);
     int increases = 0;
-    for (int i=0; i<out.size()-1; i++) {
+    for (size_t i=0; i<out.size()-1; i++) {
         increases += (out[i+1] - out[i]) > 0;
     }
     std::cout << increases << " solution1" << std::endl;
@@ -32,7 +32,7 @@ template<typename T>
 void Sol01<T>::Solution2() {
     std::vector<T> out = vTFromVVT<T>(obj);
     int increases = 0;
-    for (int i=1; i<out.size()-2; i++) {
+    for (size_t i=1; i<out.size()-2; i++) {
         int first = out[i-1] + out[i] + out[i+1];
         int second = out[i] + out[i+1] + out[i+2];
         increases += (second - first) > 0;
